# docs/swagger.yaml
basePath: /
host: 127.0.0.1:8080
info:
  contact:
    email: 244852640@qq.com
    name: 申馥源
    url: https://github.com/BIT-SFY
  description: 使用go语言实现reddit后端
  title: Go_Reddit
  version: "1.0"
paths:
  /api/v1/community:
    get:
      consumes:
      - application/json
      description: 获取社区信息
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 获取社区信息
      tags:
      - 社区模块
  /api/v1/community{id}:
    get:
      consumes:
      - application/json
      description: 根据社区id获取社区详情信息
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        type: string
      - description: 社区id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 获取社区详情信息
      tags:
      - 社区模块
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 处理用户登录请求
      parameters:
      - description: json
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 登录
      tags:
      - 用户模块
  /api/v1/posts:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        type: string
      - description: 可以为空
        in: query
        name: community_id
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子相关接口
  /api/v1/signup:
    post:
      consumes:
      - application/json
      description: 处理用户注册请求
      parameters:
      - description: json
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 注册
      tags:
      - 用户模块
swagger: "2.0"